+++ SourceryTests/Generating/StencilTemplateSpec.swift
@@ -68,7 +68,7 @@ class StencilTemplateSpec: QuickSpec {
                     "// DO NOT EDIT\n\n" +
                 "partial template content\n"
 
-                expect { try Sourcery(cacheDisabled: true).processFiles(.sources([Stubs.sourceDirectory]), usingTemplates: [templatePath], output: outputDir) }.toNot(throwError())
+                expect { try Sourcery(cacheDisabled: true).processFiles(.sources(Paths(include: [Stubs.sourceDirectory])), usingTemplates: Paths(include: [templatePath]), output: outputDir) }.toNot(throwError())
 
                 let result = (try? (outputDir + Sourcery().generatedPath(for: templatePath)).read(.utf8))
                 expect(result).to(equal(expectedResult))